# scrivia-v2-openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Scrivia API v2 - LLM Tools",
    "description": "API Scrivia v2 pour l'intégration LLM avec support complet des notes, dossiers et classeurs. Tous les endpoints supportent les slugs et UUIDs.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://scrivia.app",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Note": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "source_title": { "type": "string", "maxLength": 255 },
          "markdown_content": { "type": "string" },
          "html_content": { "type": "string" },
          "header_image": { "type": "string", "format": "uri" },
          "folder_id": { "type": "string", "format": "uuid" },
          "classeur_id": { "type": "string", "format": "uuid" },
          "user_id": { "type": "string", "format": "uuid" },
          "slug": { "type": "string" },
          "description": { "type": "string", "maxLength": 500 },
          "visibility": { "type": "string", "enum": ["private", "shared", "members", "public"] },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "source_title", "user_id", "created_at", "updated_at"]
      },
...
      "InsertContentPayload": {
        "type": "object",
        "properties": {
          "content": { "type": "string", "minLength": 1 },
          "position": { "type": "integer", "minimum": 0 }
        },
        "required": ["content", "position"]
      },
      "MergeNotePayload": {
        "type": "object",
        "properties": {
          "targetNoteId": { "type": "string", "format": "uuid" },
          "mergeStrategy": { "type": "string", "enum": ["append", "prepend", "replace"] }
        },
        "required": ["targetNoteId", "mergeStrategy"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "details": { "type": "array", "items": { "type": "string" } },
          "status": { "type": "integer" }
        },
        "required": ["error"]
      }
    }
  },
  "paths": {
...
    "/api/v2/note/{ref}/insert": {
      "post": {
        "operationId": "insertContentToNote",
        "summary": "Insérer du contenu à une position spécifique",
        "description": "Insérer du contenu markdown à une position spécifique dans la note",
        "tags": ["Notes"],
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID ou slug de la note"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/InsertContentPayload" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contenu inséré avec succès",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Note" }
              }
            }
          },
          "400": { "$ref": "#/components/schemas/ErrorResponse" },
          "404": { "$ref": "#/components/schemas/ErrorResponse" }
        }
      }
    },
...
    "/api/v2/note/{ref}/merge": {
      "post": {
        "operationId": "mergeNote",
        "summary": "Fusionner des notes",
        "description": "Fusionner le contenu d'une note avec une autre note selon une stratégie spécifique",
        "tags": ["Notes"],
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID ou slug de la note source"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MergeNotePayload" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notes fusionnées avec succès",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Note" }
              }
            }
          },
          "400": { "$ref": "#/components/schemas/ErrorResponse" },
          "404": { "$ref": "#/components/schemas/ErrorResponse" }
        }
      }
    },
...
    "/api/v2/note/{ref}/insights": {
      "get": {
        "operationId": "getNoteInsights",
        "summary": "Récupérer les insights d'une note",
        "description": "Récupérer les analyses et insights générés automatiquement pour une note",
        "tags": ["Notes"],
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID ou slug de la note"
          }
        ],
        "responses": {
          "200": {
            "description": "Insights récupérés avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "insights": { "type": "array", "items": { "type": "string" } }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/schemas/ErrorResponse" }
        }
      }
    },
...
  }
} 